seperate reducers to each campaigns sort
